src/backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
docker==7.0.0
easyocr==1.7.1
effdet==0.4.1
elastic-transport==8.13.0
elasticsearch==8.13.1
emoji==2.11.0
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gevent==24.2.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.43
gradio==4.28.3
gradio_client==0.16.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.34.1
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.3.6
langdetect==1.0.9
layoutparser==0.3.4
lazy_loader==0.4
lxml==5.2.1
magika==0.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.17.1
openai==1.16.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.1
packaging==24.0
pandas==2.2.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.0b1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.28.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-rapidjson==1.16
pytz==2024.1
pyvespa==0.40.0
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.7.0
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.2
safetensors==0.4.2
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tifffile==2024.4.24
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
tritonclient==2.45.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured==0.10.19
unstructured-inference==0.6.6
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.3
pymongo==4.7.2
python-pptx==0.6.23
azure-storage-blob==12.20.0
scikit-learn==1.5.0
mlflow==2.13.1
python-docx==1.1.2
openpyxl==3.1.2
xlrd==2.0.1
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
pyodbc==5.1.0